+++ components.php
@@ -1,13 +1,7 @@
-<?php $markup_section = "components"; ?>
-	<?php include('./template/head.php'); ?>
-	<?php include('./template/nav.php'); ?>
-		<div class="sg-body sg-container">
-
-			<div class="sg-base-styles">
-				<h1 class="sg-h1">Components</h1>
-				<?php $sg->getMarkupHtml($markup_section); ?>
-			</div><!--/.sg-base-styles-->
-
-			</div><!--/.sg-body-->
-
-	<?php include('./template/foot.php'); ?>
+<?php
+require_once("lib/Conspectus.php");
+$con = new Conspectus();
+$markup_section = "components";
+$section_title =  "Components";
+?>
+<?php include('./template/page.php'); ?>