+++ js/conspectus.js
@@ -1,0 +1,4 @@
+$( document ).ready(function() {
+  // initialize pretty print
+  prettyPrint();
+});