+++ template/foot.php
@@ -1,4 +1,13 @@
-<script src="js/sg-plugins.js"></script>
-<script src="js/sg-scripts.js"></script>
+<div class="con-footer">
+  <div class="container">
+    <strong><?= $con->getTitle() ?></strong> v.<?= $con->getVersion() ?>
+  </div>
+</div>
+  <!-- JAVASCRIPT -->
+  <script src="//code.jquery.com/jquery-1.11.1.min.js"></script>
+  <script src="//code.jquery.com/jquery-migrate-1.2.1.min.js"></script>
+  <script>window.jQuery || document.write('<script src="js/jquery-1.11.1.min.js"><\/script><script src="js/jquery-migrate-1.2.1.min.js"><\/script>')</script>
+  <script src="js/prettify.js"></script>
+  <script src="js/conspectus.js"></script>
 </body>
 </html>